Code constructor switched to double precision for channel parameters

+++ include/polarcode/arrayfuncs.h
@@ -18,7 +18,7 @@ private:
 	int partitionDescending(int lo, int hi);
 
 public:
-	float *data; ///< Pointer to memory that is to be sorted.
+	double *data; ///< Pointer to memory that is to be sorted.
 	std::vector<int> permuted; ///< Permutation vector.
 	int size; ///< Number of elements to be sorted.
 	trackingSorter();
@@ -27,21 +27,21 @@ public:
 	 * \brief Create and initialize a sorter.
 	 * \param arr The vector to be sorted.
 	 */
-	trackingSorter(std::vector<float> &arr);
+	trackingSorter(std::vector<double> &arr);
 	~trackingSorter();
 
 	/*!
 	 * \brief Set the vector to be sorted and initialize the permutation vector.
 	 * \param arr The vector to be sorted.
 	 */
-	void set(std::vector<float> &arr);
+	void set(std::vector<double> &arr);
 
 	/*!
 	 * \brief Set a subset of a vector to be sorted and initialize the permutation vector.
 	 * \param arr The vector to be sorted.
 	 * \param size The number of elements to be taken from the beginning of the vector.
 	 */
-	void set(std::vector<float> &arr, int size);
+	void set(std::vector<double> &arr, int size);
 
 	/*!
 	 * \brief Clear the contents of this object.
@@ -79,14 +79,14 @@ public:
 	 * \param n The number of highest elements to find and sort.
 	 */
 	void partialSortDescending(int n);
-	
+
 	/*!
 	 * \brief Quickly find indices of n lowest values.
 	 * \param data The vector that will be partialliy sorted.
 	 * \param size Size of that vector.
 	 * \param n Number of elements to sort.
 	 */
-	void simplePartialSort(float *data, int size, int n);
+	void simplePartialSort(double *data, int size, int n);
 
 	/*!
 	 * \brief Quickly find indices of n highest values.
@@ -94,7 +94,7 @@ public:
 	 * \param size Size of that vector.
 	 * \param n Number of elements to sort.
 	 */
-	void simplePartialSortDescending(float *data, int size, int n);
+	void simplePartialSortDescending(double *data, int size, int n);
 };
 
 //TODO: Clean up the following mess of partial sorting functions.
